Add a better pattern of name of the files
This is a next step of the previous refactor. Before the usecases
was inside the entity. So, registering was a file of
"/transaction/usecases/registering.go". Now, all usecases are inside
the usecases package I need to given the context of the registeing.

Is it registering of what? So, now the file starts if transaction_
to be more clear.