Add funding handler test
Introducing tests in http handlers.

A handler_mock_test.go file was created. This file has functions to
mock request and response to be used inside the tests.

In the funding test I noticed that the status code returned was 200
when the correct status code of an funding creation is 201.