Add retrieve transaction use case
To be possible to do it, the TransactionRepository interface
has a new method called FindByID. And in the memory repository was
created a fake implementations just to compile the application.

The next step of this task is to create the infrastructure of the
application. That is: The correct implementation of the memory
repository and the handlers of the application.